requirements.txt
environs==11.0.0
Flask==3.0.3
Werkzeug==3.0.4

googleapis-common-protos==1.63.0
grpcio==1.48.1
grpcio-health-checking==1.48.1
grpcio-reflection==1.48.1
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-exporter-prometheus==0.35b0
opentelemetry-exporter-zipkin-proto-http==1.14.0
opentelemetry-instrumentation-grpc==0.35b0
opentelemetry-propagator-b3==1.19.0
prometheus-client==0.20.0
protobuf==3.20.3
python-logging-loki==0.3.1

accelbyte-py-sdk
bitarray
httpx[http2]
mmh3
PyJWT[crypto]
PyYAML
requests
websockets